src/main/java/longest/path/dag/Model/Vertex.java
package longest.path.dag.Model;

import java.util.Objects;

public class Vertex {
    private long id;  // unique identifier for the vertex

    public Vertex(long id) {
        this.id = id;
    }

    public long getId() {
        return id;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) return true;
        if (obj == null || getClass() != obj.getClass()) return false;
        Vertex vertex = (Vertex) obj;
        return id == vertex.id;
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }
}

